Improved Dynamic Interface + Hinge BCs (in progress):
- Dynamic Interface (Python):
	+ improved options for input (static, dynamic, distributed and user defined loads)
	+ new methods for forced velocities
	+ save/read dynamics related output
	+ post-processing module added

- Dynamic Interface (Fortran)
	+ deleted some unecessary input
	+ commented printing options (giving error for large files)

- Hinge BCs: 
This mainly affects the Fortran part of the code. Previously implemented BCs are currently working, but hinged BCs are still a work in progress. 

The original code uses a strong enforcement of BCs (clamped BCs only + free-end allowed), meaning that the system size depends to the BCs. The current effort has been put in trying to modify the system dimension to account for the different number of  dofs and eq.s to solve (cbeam3_asbly.f90) when hinged BCs are used.

This approach, however, requires changes several parts of ther code (assembly/utility/ordering methods - e.g. lib_fem/fem_ma2v  - which assume that for each node you can eith have 6 or 0 dofs associated) and is probably not the most robust way to procede.

The current code( with the attempt to change the system size) is, therefore, committed but the changes affecting the system size will be retrieved and the implementation of hinge BCs will be done trying not to change the criteria determining the system size (weak enforcement).