Fix gradient step reduction bug + mpatplotlib 1.4.3 compatibility update
Details:

Step reduction bug fix:
Probably with the intro of FD manually computed gradient, the step reduction would fail as the old design variables values would not be retained. Values of deisgn are now saved in h5 file, and they are read if necessary.
- xbcompFD: design values are saved into DesignVal and saved in h5. If crash occurs, they are read from the old file
- lib.save, lib.read: added 'DesignVal'

- WARNING: 
a. during FD the method would fail, as the step reduction will be computed in respect to the previous run, not the current design. For FD, in general, using the method makes no sense.

Matplotlib:
- update installations instructions
- lib.plot.share: new standard for plotting introduced