Spherical Joint working (bug fixed)
Summary: spherical joints fully working and bug in force influence matrix for rigid body dynamics fixed. When a spherical joint is used, the beam is shifted such that the joint coincides with the origin of the FoR A. The BCs are unchanged in respect to the past. A bug in the force influence coefficient matrix was found: whenever an applied force would precede the clamp node, the related influence coefficients would be wrong. 


- xbeam_asbly.f90: bug in Frigid calculation fixed

- beamgeom.py/xbcompFD.py: methods to shift the beam position such that the clamp coincides with the origin of FoR A. This allows not to change the BCs implementation in the xbeam_solv