Merge Checking: dynamic structural solution (called from BeamSolver wrapper) working
The following set of cases have been run:
	- pendulum
	- arbitrary spherical joint position
	- free-flying beam with arbitrary clamp point

Note that the interface to run these cases requires to compile not only the BeamLib.so library in the optimiser area for solution BUT also (1) the BeamLib.so library in the SHARPy wrapper area and (b) the f2py codes. This because in the postprocessing, the PyBeams/Utils/BeamLib.py module is imported. 


List of corrections:

cbeam3_solv.f90: 
	- restored use lib_solv to check convergence
	- use lu_solve also in the "optimal_control" copy of the solvers (remove lapack_sparse calls)
	- identity between cbeam3_nlndyn_solv and cbeam3_nlndyn_solv_opt_control verified

xbeam3_solv.f90:
	- use lu_solve also in the "optimal_control" copy of the solvers (remove lapack_sparse calls)
	- identity between cbeam3_nlndyn_solv and cbeam3_nlndyn_solv_opt_control verified
	- cleaned-up
	- bug fix: initial inversion to find initial acceleration restored

runf2py27.sh: added to allow to compile lib_fem with python2.7 standard.