FD in parallel working:
Allocation of cpu usage optimised to use as many cpus as possible for each
perturbed run. Performance optimisation implemented only for when the 
amount of processors is larger then the number of design variables but is not guaranteed to reduce the total run time.

Other:

PyBeam.Utils.DerivedTypes: 
	- added methods to set/delete the ctypes options. The methods are not
necessary to run a pool of solvers contemporary (as in python 2.x) but have
been kept.