Major bug fixed and main flight dynamics solver rewritten.
The flight dynamic solver has been reorganised in a more logical way that facilitates the check/understanding of how inofrmations flows. A number of minor bugs were fixed (mainly variables evaluated at the wrong time-step or not updated during Newton-Raphson). A bug in the gravity loads calculation was found and fixed (loads only accounting for initial beam orientation!).


PyBeam:

- PyBeam.Utils.XbeamLib: 
	AddGravityLoads routine extended to account for current orientation of beam
- PyBeam.Utils.DerivedTypes:
	added method to Xboutput to facilitate saving


PyCoupled:

- PyCoupled.Coupled_NlnFlightDynamic_lag: 
	This module rewrites the original flight-dynamics solution (Coupled_NlnFlightDynamic). The solver has been restructured in a more logical way and the lines of code have been reduced defining some methods to implement repetitive/nonrelevant operations (initialisation and assembly of GEBM tensors, computation of UVLM solution). These have been collected in Coupled_NlnFlightDynamic_support.  
The flow of informations has also been simplified and a number of minor bugs (variables evaluated at wrong time-step) have been fixed. A major fix involved the gravity loads (XbeamLib), which were not accounting for the wing attitude in the FoR G. Also, external loads (except aerodynamics), are now evaluated at each iteration of the Newton-Raphson iteration. The aeroelastic coupling can be regulated: by default, the aerodynamic solution is computed based on the beam geometry at the first iteration of the Newton-Raphson (as per an explicit time-marching scheme).

- PyCoupled.Coupled_NlnFlightDynamic_support:
	Contains methods to initialise, assemble GEBM tensor, compute lagged solution and UVLM nodal forces.

- PyCoupled.Utils.DerivedTypesAeroelastic:
	Added parameters to regulate the aeroelastic coupling in flight dynamics solution. 


PyOpt:

- PyOpt.main: 
	PyCoupled.Coupled_NlnFlightDynamic_lag made default solver